add bpm input
