+++ NTM_M4L/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "max-client",
+  "version": "1.0.0",
+  "description": "",
+  "main": "max-client.js",
+  "engines": {
+    "node": ">=16.x",
+    "npm": ">=7.x"
+  },
+  "scripts": {},
+  "author": "Ben Johansen <ben@benjohansen.com> (https://benjohansen.com/)",
+  "license": "MIT",
+  "dependencies": {
+    "socket.io-client": "^4.6.1"
+  }
+}
