+++ NTM_Transport/style.css
@@ -1,0 +1,27 @@
+.bang
+{
+    width: 100px;
+}
+
+#recieveBang
+{
+    pointer-events: none;
+}
+
+#sendBang:active
+{
+    background-color: blue;
+}
+
+@media screen and (max-width: 500px)
+{
+    .bang
+    {
+        width: 50%;
+        font-size: 1.5rem;
+    }
+    #recieveBang
+    {
+        float: left;
+    }
+}
