--- old-index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Netronome Transport</title>
-</head>
-<body>
-    <p>Netronome Transport</p>
-
-    <label for="bpm">bpm: </label>
-    <input type="number" id="bpm" name="bpm" min="1" max="999" value="120" class="BPM">
-
-    <script src="./NTM_Transport/script.js"></script>
-</body>
-</html>
